COMP: Unusually formatted code is error prone
Commonly mis-formatted code structures identified.  The code
was correct, but very easy to mis-interpret.

Removing this nit-picky compiler warning removes noise from
the dashboard, and inproves code readability for new developers.
The current formatting is less compact, but it provides indentation
visual queues for identifying implicit blocks for if statements.

Previous code of:
```c++
if(test) statment(with,var,arguments; break
```

is now more visually indicative that the `break` is not part of the if
block.

```c++
if(test)
  statement(with,var,arguments);
break;
```