COMP: Remove 'Default case is not handled' warnings.
Explicitly handling the default case in a switch statement
is considered a best practice.

The clang-tidy tool performs analysis and attempts to find
code uses where common errors are present.  It does not mean
that the code is wrong, but that the particular style of
coding used is often a source of unexpected bugs.