make test data download optional
Requested by @yarikoptic for the purposes of debian packaging.
With this commit, when the cmake option DOWNLOAD_TEST_DATA is set to OFF,
the test data will not be fetched.
In conjunction the NEEDS_DATA label can be used to filter out dependent
tests i.e. ctest -LE NEEDS_DATA