tidy installation pattern.
Tidies support for namespaced targets as part of superbuild, now adding
the aliased library within the add_nifti_library function.

All functionality required for ITK to drive the build now working. i.e.
full supports the NIFTI_INSTALL_NO... vars.

Manages all header file installation using target properties. Either
PUBLIC_HEADER or PRIVATE_HEADER can be used depending on whether the
header should be installed.

For cmake >=3.13, we now aggregate the list of targets to be
installed and install them all in the main cmake file to reduce
amount of code required and to reduce the chance of errors of
installation directory names etc.

For older cmake versions the above pattern for target installation is
not possible because targets are not scoped beyond the directory in
which they are defined. The approach taken for these cmake versions is
to use the `install_nifti_target` function in the targets cmake file
(after all target properties etc. have been set). These lines can be
dropped in the future.